+++ tests/functional_tests/workgroup_primitives.cpp
@@ -40,7 +40,6 @@ __global__ void WorkGroupPrimitiveTest(int loop, int skip,
                                       ShmemContextType ctx_type) {
   __shared__ rocshmem_ctx_t ctx;
   int wg_id = get_flat_grid_id();
-  rocshmem_wg_init();
   rocshmem_wg_ctx_create(ctx_type, &ctx);
 
   // Calculate start index for each work group
@@ -82,7 +81,6 @@ __global__ void WorkGroupPrimitiveTest(int loop, int skip,
   }
 
   rocshmem_wg_ctx_destroy(&ctx);
-  rocshmem_wg_finalize();
 }
 
 /******************************************************************************